Inject decorator over check_in and new_checkin
